+++ apps/crews/src/Agent/Agent.prompts.ts
@@ -1,0 +1,26 @@
+/**
+ * Agent
+ */
+export const AgentSystemPrompt = (name: string, role: string, systemPrompt?: string) =>
+  `You are: ${name}
+Your role is: ${role}
+${systemPrompt ? `\n${systemPrompt}\n` : ""}
+Be brief and concise!`;
+
+export const TaskPrompt = (task: string, availableTools: string) =>
+  `# Your Task is:\n${task}\n\n# These are the tools available for you to use:\n${availableTools}\n\nBegin! This is VERY important to you, use any tools available and give your best final answer!`;
+
+/**
+ * Manager
+ */
+export const ManagerSystemPrompt = (
+  coworkers: string,
+) => `You are a skilled manager, known for your ability to delegate work to the right people, and to ask the right questions to get the best out of your team.
+Don't perform the tasks by yourself.
+MAKE SURE to only respond with the answers of your team members. Don't share your own answers or evaluations.
+
+Your team consists of:
+${coworkers}
+`;
+
+export const ManagerRole = "Manage the team to complete the task in the best way possible";