+++ apps/ui/app/components/traces/MessagesView/styles.tsx
@@ -8,10 +8,10 @@ export const Content: FC<{ children: ReactNode }> = ({ children }) => {
   return <div className={"bg-secondary dark:bg-secondary/25 whitespace-pre-wrap rounded-xl px-8 py-4"}>{children}</div>;
 };
 
-export const Container: FC<{ children: ReactNode }> = ({ children }) => {
-  return <div className={"whitespace-pre-wrap text-wrap"}>{children}</div>;
+export const MessagesContainer: FC<{ children: ReactNode }> = ({ children }) => {
+  return <div className={"flex flex-col gap-4"}>{children}</div>;
 };
 
-export const MessageContainer: FC<{ children: ReactNode }> = ({ children }) => {
-  return <div className={"flex flex-col gap-4"}>{children}</div>;
+export const Message: FC<{ children: ReactNode }> = ({ children }) => {
+  return <div className={"flex flex-col gap-1"}>{children}</div>;
 };