+++ packages/config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @montelo/config
 
+## 1.1.65
+
+### Patch Changes
+
+- Crews
+
+## 1.1.64
+
+### Patch Changes
+
+- fix
+
 ## 1.1.63
 
 ### Patch Changes