+++ packages/promptfile/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @montelo/promptfile
 
+## 1.1.65
+
+### Patch Changes
+
+- Crews
+- Updated dependencies
+  - @montelo/common@1.1.65
+
+## 1.1.64
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @montelo/common@1.1.64
+
 ## 1.1.63
 
 ### Patch Changes